feat(config): add a configurable `jira_token`
Customize the token used for the footer containing Jira issues by adding a `jira_token` key to the .cz.yaml|json|toml config file.

The token must be acceptable per the conventional commit spec and cannot be equal to any variation of "BREAKING CHANGE". If not
specified, the token is set to "Jira: "

Jira: XX-02